Return MSFList, not plain list, from form field
