+++ multiselectfield/forms/fields.py
@@ -16,7 +16,7 @@
 
 from django import forms
 
-from ..utils import get_max_length
+from ..utils import MSFList, get_max_length
 from ..validators import MaxValueMultiFieldValidator, MinChoicesValidator, MaxChoicesValidator
 
 
@@ -27,6 +27,7 @@ class MultiSelectFormField(forms.MultipleChoiceField):
         self.min_choices = kwargs.pop('min_choices', None)
         self.max_choices = kwargs.pop('max_choices', None)
         self.max_length = kwargs.pop('max_length', None)
+        self.flat_choices = kwargs.pop('flat_choices')
         super(MultiSelectFormField, self).__init__(*args, **kwargs)
         self.max_length = get_max_length(self.choices, self.max_length)
         self.validators.append(MaxValueMultiFieldValidator(self.max_length))
@@ -34,3 +35,6 @@ class MultiSelectFormField(forms.MultipleChoiceField):
             self.validators.append(MaxChoicesValidator(self.max_choices))
         if self.min_choices is not None:
             self.validators.append(MinChoicesValidator(self.min_choices))
+
+    def to_python(self, value):
+        return MSFList(dict(self.flat_choices), super(MultiSelectFormField, self).to_python(value))
