+++ multiselectfield/utils.py
@@ -25,6 +25,21 @@ else:
     string_type = string
 
 
+class MSFList(list):
+
+    def __init__(self, choices, *args, **kwargs):
+        self.choices = choices
+        super(MSFList, self).__init__(*args, **kwargs)
+
+    def __str__(msgl):
+        msg_list = [msgl.choices.get(int(i)) if i.isdigit() else msgl.choices.get(i) for i in msgl]
+        return u', '.join([string_type(s) for s in msg_list])
+
+    if sys.version_info < (3,):
+        def __unicode__(self, msgl):
+            return self.__str__(msgl)
+
+
 def get_max_length(choices, max_length, default=200):
     if max_length is None:
         if choices:
